requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.7.0
attrs==24.2.0
awscli==1.36.19
boto3==1.35.78
botocore==1.35.78
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.7.7
deprecation==2.1.0
docutils==0.16
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
marshmallow==3.23.1
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
packaging==24.2
propcache==0.2.1
pyasn1==0.6.1
pydub==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
